Extract home fucntion to home module
