+++ src/index.js
@@ -1,36 +1,9 @@
+import homeFunc from './home.js'
+
 const content = document.getElementById("content");
 const homeDiv = document.createElement("div");
 const menuDiv = document.createElement("div");
 const contactDiv = document.createElement("div");
-
-const homeFunc = () => {
-  const homeDivOne = document.createElement("div");
-  const homeDivTwo = document.createElement("div");
-  const heading = document.createElement("h1");
-  const paragragh = document.createElement("p");
-
-  homeDiv.id = "home";
-  homeDiv.classList.add(
-    "top-bg",
-    "homeContainer"
-  );
-  homeDivOne.classList.add("homeContent",);
-  homeDivTwo.classList.add("homeBox");
-  heading.classList.add("homeTitle");
-  paragragh.classList.add("homePg");
-  
-  heading.textContent = "Welcome to Alorr!";
-  paragragh.textContent =
-    "Lorem ipsum dolor sit amet consectetur, adipisicing elit. Quam cupiditate, ipsum saepe odio sint neque facilis quod culpa animi, dolorem consequuntur consequatur. Ea, ex voluptatum! Quam, ratione dolorum? In, reiciendis nihil. Minima saepe iure ut ex praesentium. Quidem esse repellendus molestias similique expedita illum?";
-
-  homeDivTwo.appendChild(heading);
-  homeDivTwo.appendChild(paragragh);
-  homeDivOne.appendChild(homeDivTwo);
-  homeDiv.appendChild(homeDivOne);
-
-
-  return homeDiv;
-}
 
 const menuFunc = () => {
   const heading = document.createElement("h2");
@@ -46,9 +19,11 @@ const menuFunc = () => {
   const boxFourContainer = document.createElement("div");
   const boxImgOne = document.createElement("img");
   const boxImgTwo = document.createElement("img");
-  const title = document.createElement("h4");
+  const titleOne = document.createElement("h4");
+  const titleTwo = document.createElement("h4");
   const paragraph = document.createElement("p");
   const paragraphText = "Doloremque omnis exercitationem nulla ratione perferendis, quam, incidunt earum assumenda eius.";
+
   menuDiv.id = "menu";
   menuDiv.classList.add(
     "menu-bg",
@@ -58,24 +33,62 @@ const menuFunc = () => {
   heading.classList.add("menuTitle");
   container.classList.add("menu-v");
   gridBox.classList.add("menuGrid");
-  boxOneContainer.classList.add("grid-iv")
+
   boxOne.classList.add("gridItem");
-  boxImgOne.src = "../assets/img/6.jpg";
-  boxImgOne.alt = "Peppered Rice";
+  boxTwo.classList.add("gridItem");
+  boxThree.classList.add("gridItem");
+  boxFour.classList.add("gridItem");
+
   boxImgOne.classList.add("gridImage");
-  title.textContent = "Peppered Rice";
-  title.classList.add("gridTitle");
+  boxImgTwo.classList.add("gridImage");
+
+  boxOneContainer.classList.add("grid-iv");
+  boxTwoContainer.classList.add("grid-iv");
+  boxThreeContainer.classList.add("grid-iv");
+  boxFourContainer.classList.add("grid-iv");
+
+  titleOne.textContent = "Peppered Rice";
+  titleOne.classList.add("gridTitle");
+
+  titleTwo.textContent = "Huevos Rotos";
+  titleTwo.classList.add("gridTitle");
+
   paragraph.textContent = paragraphText;
   paragraph.classList.add("gridPg");
+
+  boxImgOne.src = "../assets/img/6.jpg";
+  boxImgTwo.src = "../assets/img/11.png";
+
+  boxImgOne.alt = "Peppered Rice";
+  boxImgTwo.alt = "Huevos Rotos";
 
   menuDiv.appendChild(heading);
   menuDiv.appendChild(container);
   container.appendChild(gridBox);
+
   gridBox.appendChild(boxOne);
   boxOne.appendChild(boxImgOne);
   boxOne.appendChild(boxOneContainer);
-  boxOneContainer.appendChild(title);
+  boxOneContainer.appendChild(titleOne);
   boxOneContainer.appendChild(paragraph);
+
+  // gridBox.appendChild(boxTwo);
+  // boxTwo.appendChild(boxImgTwo);
+  // boxTwo.appendChild(boxTwoContainer);
+  // boxTwoContainer.appendChild(titleTwo);
+  // boxTwoContainer.appendChild(paragraph);
+
+  // gridBox.appendChild(boxThree);
+  // boxThree.appendChild(boxImgOne);
+  // boxThree.appendChild(boxThreeContainer);
+  // boxThreeContainer.appendChild(titleTwo);
+  // boxThreeContainer.appendChild(paragraph);
+  
+  // gridBox.appendChild(boxFour);
+  // boxFour.appendChild(boxImgTwo);
+  // boxFour.appendChild(boxFourContainer);  
+  // boxFourContainer.appendChild(titleOne);
+  // boxFourContainer.appendChild(paragraph);
 
   return menuDiv
 }
@@ -114,7 +127,6 @@ const contactFunc = () => {
 }
 
 const showContent = () => {
-
   content.appendChild(homeFunc(homeDiv));
   content.appendChild(menuFunc(menuDiv));
   content.appendChild(contactFunc(contactDiv));
